[UPDATE] Add evaluation metrics

+++ magnet/options/test.py
@@ -6,7 +6,7 @@ class TestOptions(BaseOptions):
         super().__init__()
         parser = self.parser
         parser.add_argument('--sub_batch_size', required=True, type=int, help='batch size for patches')
-        parser.add_argument('--n_points', required=True, type=int, help='number of points to be refined at each stage')
+        parser.add_argument('--n_points', required=True, type=float, help='number of points to be refined at each stage')
         parser.add_argument('--smooth_kernel', default=16, type=int, help='kernel size to Gaussian blur error scores')
         self.parser = parser
